+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and as of version 1.0.0, follows semantic versioning.
 
 ## [Unreleased]
+## Added
+  * Shesmu input link to the `status` page to view JSON that's passed to ETL
 
 ## [210412-1508] - 2021-04-12
 ## Changed